requirements.txt
openai>=1.12.0
pydantic>=2.5.0
chromadb>=0.4.22
sqlalchemy>=2.0.25
python-dotenv>=1.0.0
typing-extensions>=4.9.0

